requirements.txt
torch==2.0.1
torchvision==0.15.2
pandas==1.5.3
numpy==1.24.0
matplotlib==3.6.3
scikit-learn==1.1.3
biopython
Pillow==9.2.0